Moved Disc class to Disc.py
